feat: consumer host services
