+++ src/hooks/index.ts
@@ -19,7 +19,6 @@ export * from './useIsIndexer';
 export * from './useSortedIndexerDeployments';
 export * from './useNetworkClient';
 export * from './useSwapData';
-export * from './useIndexerFlexPlans';
 export * from './useInitContracts';
 export * from './useEra';
 export * from './useStudio';
